w near split into pub and internal functions
